refactor(kafka): renamed all kafka notification services & moved to external package
