+++ src/main/kotlin/org/careerseekers/csmailservice/services/interfaces/IDirectionDocsTasksProcessingService.kt
@@ -3,6 +3,6 @@ package org.careerseekers.csmailservice.services.interfaces
 import org.careerseekers.csmailservice.dto.DirectionDocumentsTask
 import org.careerseekers.csmailservice.enums.DirectionDocsEventTypes
 
-interface IDirectionDocsTasksProcessingService : KafkaMessageHandler<String, DirectionDocumentsTask> {
+interface IDirectionDocsTasksProcessingService : IKafkaMessageHandler<String, DirectionDocumentsTask> {
     val eventType: DirectionDocsEventTypes
 }
