+++ src/main/kotlin/org/careerseekers/csmailservice/services/interfaces/IEmailNotificationProcessingService.kt
@@ -4,7 +4,7 @@ import org.careerseekers.csmailservice.dto.EmailSendingTaskDto
 import org.careerseekers.csmailservice.enums.MailEventTypes
 import org.springframework.mail.javamail.JavaMailSender
 
-interface EmailProcessingService : KafkaMessageHandler<String, EmailSendingTaskDto> {
+interface IEmailNotificationProcessingService : IKafkaMessageHandler<String, EmailSendingTaskDto> {
     val mailer: JavaMailSender
     val eventType: MailEventTypes
 }
