+++ src/main/kotlin/org/careerseekers/csmailservice/services/interfaces/IKafkaMessageHandler.kt
@@ -2,7 +2,7 @@ package org.careerseekers.csmailservice.services.interfaces
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 
-interface KafkaMessageHandler<T, K> {
+interface IKafkaMessageHandler<T, K> {
 
     fun handle(record: ConsumerRecord<T, K>)
 }
