+++ src/main/kotlin/org/careerseekers/csmailservice/services/kafka/consumers/KafkaEmailSendingConsumer.kt
@@ -2,14 +2,14 @@ package org.careerseekers.csmailservice.services.kafka.consumers
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.careerseekers.csmailservice.dto.EmailSendingTaskDto
-import org.careerseekers.csmailservice.services.interfaces.EmailProcessingService
+import org.careerseekers.csmailservice.services.interfaces.IEmailNotificationProcessingService
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.kafka.support.Acknowledgment
 import org.springframework.stereotype.Service
 
 @Service
 class KafkaEmailSendingConsumer(
-    private val emailProcessingServices: List<EmailProcessingService>
+    private val emailNotificationProcessingServices: List<IEmailNotificationProcessingService>
 ) : CustomKafkaConsumer<String, EmailSendingTaskDto> {
 
     @KafkaListener(
@@ -20,7 +20,7 @@ class KafkaEmailSendingConsumer(
         consumerRecord: ConsumerRecord<String, EmailSendingTaskDto>,
         acknowledgment: Acknowledgment
     ) {
-        for (service in emailProcessingServices) {
+        for (service in emailNotificationProcessingServices) {
             if (service.eventType == consumerRecord.value().eventType) {
                 service.handle(consumerRecord)
                 acknowledgment.acknowledge()
