+++ src/main/kotlin/org/careerseekers/csmailservice/services/kafka/consumers/KafkaPlatformCreationConsumer.kt
@@ -2,14 +2,14 @@ package org.careerseekers.csmailservice.services.kafka.consumers
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.careerseekers.csmailservice.dto.PlatformCreationDto
-import org.careerseekers.csmailservice.services.PlatformCreationService
+import org.careerseekers.csmailservice.services.notifications.PlatformCreationNotificationService
 import org.springframework.kafka.annotation.KafkaListener
 import org.springframework.kafka.support.Acknowledgment
 import org.springframework.stereotype.Service
 
 @Service
 class KafkaPlatformCreationConsumer(
-    private val platformCreationService: PlatformCreationService
+    private val platformCreationNotificationService: PlatformCreationNotificationService
 ) : CustomKafkaConsumer<String, PlatformCreationDto> {
 
     @KafkaListener(
@@ -20,7 +20,7 @@ class KafkaPlatformCreationConsumer(
         consumerRecord: ConsumerRecord<String, PlatformCreationDto>,
         acknowledgment: Acknowledgment
     ) {
-        platformCreationService.handle(consumerRecord)
+        platformCreationNotificationService.handle(consumerRecord)
         acknowledgment.acknowledge()
     }
 }
