+++ src/main/kotlin/org/careerseekers/csmailservice/services/notifications/DirectionCreationNotificationService.kt
@@ -1,33 +1,30 @@
-package org.careerseekers.csmailservice.services
+package org.careerseekers.csmailservice.services.notifications
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.careerseekers.csmailservice.config.MailProperties
-import org.careerseekers.csmailservice.dto.DirectionDocumentsTask
-import org.careerseekers.csmailservice.enums.DirectionDocsEventTypes
-import org.careerseekers.csmailservice.services.interfaces.IDirectionDocsTasksProcessingService
+import org.careerseekers.csmailservice.dto.DirectionCreation
+import org.careerseekers.csmailservice.services.interfaces.IKafkaMessageHandler
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.mail.SimpleMailMessage
 import org.springframework.mail.javamail.JavaMailSender
 import org.springframework.stereotype.Service
 
 @Service
-class DirectionDocsCreationService(
+class DirectionCreationNotificationService(
     @param:Qualifier("productionMailSender") private val mailer: JavaMailSender,
     private val mailProperties: MailProperties,
-) : IDirectionDocsTasksProcessingService {
+) : IKafkaMessageHandler<String, DirectionCreation> {
 
-    override val eventType = DirectionDocsEventTypes.CREATION
-
-    override fun handle(record: ConsumerRecord<String, DirectionDocumentsTask>) {
+    override fun handle(record: ConsumerRecord<String, DirectionCreation>) {
         val message = record.value()
 
         SimpleMailMessage().apply {
             from = mailProperties.productionMail.username
-            setTo(message.tutor.email)
-            subject = "Искатели Профессий | Загрузка нового документа"
+            setTo(message.expert.email)
+            subject = "Искатели Профессий | Создание новой компетенции"
             text = """
-            Уважаемый(-ая) ${message.tutor.lastName} ${message.tutor.firstName} ${message.tutor.patronymic}!   
-            Сообщаем Вам, что ${message.expert.lastName} ${message.expert.firstName} ${message.expert.patronymic} добавил новый документ для компетенции ${message.directionName}, в возрастной категории ${message.ageCategory}.
+            Уважаемый(-ая) ${message.expert.lastName} ${message.expert.firstName} ${message.expert.patronymic}!   
+            Сообщаем, что Вам добавлена новая компетенция — ${message.name}.
 
             Более подробную информацию Вы можете увидеть в своем Личном кабинете.
