+++ src/main/kotlin/org/careerseekers/csmailservice/services/notifications/ExpertRegistrationNotificationServiceNotificationI.kt
@@ -1,4 +1,4 @@
-package org.careerseekers.csmailservice.services
+package org.careerseekers.csmailservice.services.notifications
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.careerseekers.csmailservice.cache.TemporaryPasswordsCache
@@ -6,18 +6,18 @@ import org.careerseekers.csmailservice.config.MailProperties
 import org.careerseekers.csmailservice.dto.EmailSendingTaskDto
 import org.careerseekers.csmailservice.enums.MailEventTypes
 import org.careerseekers.csmailservice.exceptions.BadRequestException
-import org.careerseekers.csmailservice.services.interfaces.EmailProcessingService
+import org.careerseekers.csmailservice.services.interfaces.IEmailNotificationProcessingService
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.mail.SimpleMailMessage
 import org.springframework.mail.javamail.JavaMailSender
 import org.springframework.stereotype.Service
 
 @Service
-class ExpertRegistrationEmailService(
+class ExpertRegistrationNotificationServiceNotificationI(
     @param:Qualifier("productionMailSender") override val mailer: JavaMailSender,
     private val temporaryPasswordsCache: TemporaryPasswordsCache,
     private val mailProperties: MailProperties,
-) : EmailProcessingService {
+) : IEmailNotificationProcessingService {
     override val eventType = MailEventTypes.EXPERT_REGISTRATION
 
     override fun handle(record: ConsumerRecord<String, EmailSendingTaskDto>) {
