+++ src/main/kotlin/org/careerseekers/csmailservice/services/notifications/MentorAndUserRegistrationNotificationServiceNotificationI.kt
@@ -1,20 +1,20 @@
-package org.careerseekers.csmailservice.services
+package org.careerseekers.csmailservice.services.notifications
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.careerseekers.csmailservice.config.MailProperties
 import org.careerseekers.csmailservice.dto.EmailSendingTaskDto
 import org.careerseekers.csmailservice.enums.MailEventTypes
-import org.careerseekers.csmailservice.services.interfaces.EmailProcessingService
+import org.careerseekers.csmailservice.services.interfaces.IEmailNotificationProcessingService
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.mail.SimpleMailMessage
 import org.springframework.mail.javamail.JavaMailSender
 import org.springframework.stereotype.Service
 
 @Service
-class MentorAndUserRegistrationEmailService(
+class MentorAndUserRegistrationNotificationServiceNotificationI(
     @param:Qualifier("productionMailSender") override val mailer: JavaMailSender,
     private val mailProperties: MailProperties,
-) : EmailProcessingService {
+) : IEmailNotificationProcessingService {
 
     override val eventType = MailEventTypes.MENTOR_AND_USER_REGISTRATION
